requirements.txt.txt
# Web 框架
streamlit>=1.28.0

# LLM 接口管理
litellm>=1.20.0

# 可视化
plotly>=5.18.0

# 数据处理
pandas>=2.0.0

# 环境变量
python-dotenv>=1.0.0

# 数据验证
pydantic>=2.0.0

# 数学计算
numpy>=1.24.0

# 测试 (可选)
pytest>=7.4.0
